refactor: extract test utils
