+++ frontend/src/tests/lib/modals/sns/SnsStakeMaturityModal.spec.ts
@@ -9,6 +9,7 @@ import { fireEvent, waitFor, type RenderResult } from "@testing-library/svelte";
 import type { SvelteComponent } from "svelte";
 import { renderModal } from "../../../mocks/modal.mock";
 import { mockSnsNeuron } from "../../../mocks/sns-neurons.mock";
+import { selectPercentage } from "../../../utils/neurons-modal.test-utils";
 
 jest.mock("$lib/services/sns-neurons.services", () => {
   return {
@@ -39,22 +40,6 @@ describe("SnsStakeMaturityModal", () => {
 
     expect(queryByText(formattedMaturity(mockSnsNeuron))).toBeInTheDocument();
   });
-
-  const selectPercentage = async (
-    renderResult: RenderResult<SvelteComponent>
-  ) => {
-    const { queryByTestId } = renderResult;
-    const rangeElement = queryByTestId("input-range");
-    expect(rangeElement).toBeInTheDocument();
-    rangeElement &&
-      (await fireEvent.input(rangeElement, { target: { value: 50 } }));
-
-    const selectMaturityButton = queryByTestId(
-      "select-maturity-percentage-button"
-    );
-    expect(selectMaturityButton).toBeInTheDocument();
-    selectMaturityButton && (await fireEvent.click(selectMaturityButton));
-  };
 
   it("should call stakeMaturity service on confirm click", async () => {
     const renderResult: RenderResult<SvelteComponent> =
